fiddled with animation code
animation is still very finicky, mainly due to the fact that GameSprite
uses name for image, which doesnt work if youre using frames for image
in AnimSprite. Maybe we can just try/except the image loading in
GameSprite and override it with AnimSprite to use frames. one problem as
well is that if you want to have an animation complete an action on its
animation being completed, you would have to manually do that in the
on_screen_animation codeblock, if anim.playing is not equal to True. How
should we go about this? I may put up an issue request.