added comments so code is easier to understand (ps code needs refactoring)
if theres anything else you need me to explain im glad to comment it or
explain in a later meeting, mb for the hard to read code. i think that
the code works in a weird way, and each system is too reliant on other
systems. each function/class needs to be more abstract/general/decoupled
so that it can work with a range of things, and won't run into issues as
often.